+++ pages/wildcard.clang_switch.php
@@ -24,8 +24,27 @@ $clang_id = (int)str_replace('clang', '', rex_be_controller::getCurrentPagePart(
 $error = '';
 $success = '';
 
+// Wenn der Platzhalter vom Admin geändert wird, muss dieser in den anderen Sprachen synchronisiert werden
+if (rex::getUser()->isAdmin() && count(rex_clang::getAll()) >= 2) {
+    \rex_extension::register('REX_FORM_SAVED', function(rex_extension_point $ep) use ($pid, $clang_id) {
+        $form = $ep->getParam('form');
+        if ($form->isEditMode()) {
+            $items = rex_sql::factory()->getArray('SELECT `id`, `wildcard` FROM ' . $form->getTablename() . ' WHERE `pid` = :pid LIMIT 2', ['pid' => $pid]);
+            if (count($items) == 1) {
+                $savedId = $items[0]['id'];
+                $savedWildcard = $items[0]['wildcard'];
+                $sql = rex_sql::factory();
+                $sql->setTable($form->getTablename());
+                $sql->setWhere('pid != :pid AND id = :id', ['pid' => $pid, 'id' => $savedId]);
+                $sql->setValue('wildcard', $savedWildcard);
+                $sql->update();
+            }
+        }
+    });
+}
+
 // ----- delete wildcard
-if ($func == 'delete' && $wildcard_id > 0) {
+if ($func == 'delete' && $wildcard_id > 0 && rex::getUser()->isAdmin()) {
     $deleteWildcard = rex_sql::factory();
     $deleteWildcard->setQuery('DELETE FROM ' . rex::getTable('sprog_wildcard') . ' WHERE id=?', [$wildcard_id]);
     $success = $this->i18n('wildcard_deleted');
@@ -58,11 +77,17 @@ if ($func == '') {
     $list->setColumnLayout('edit', ['<th class="rex-table-action" colspan="2">###VALUE###</th>', '<td class="rex-table-action">###VALUE###</td>']);
     $list->setColumnParams('edit', ['func' => 'edit', 'pid' => '###pid###']);
 
-    $list->addColumn('delete', '<i class="rex-icon rex-icon-delete"></i> ' . $this->i18n('delete'));
-    $list->setColumnLabel('delete', $this->i18n('function'));
-    $list->setColumnLayout('delete', ['', '<td class="rex-table-action">###VALUE###</td>']);
-    $list->setColumnParams('delete', ['func' => 'delete', 'wildcard_id' => '###id###']);
-    $list->addLinkAttribute('delete', 'data-confirm', $this->i18n('delete') . ' ?');
+    if (rex::getUser()->isAdmin()) {
+        $list->addColumn('delete', '<i class="rex-icon rex-icon-delete"></i> ' . $this->i18n('delete'));
+        $list->setColumnLabel('delete', $this->i18n('function'));
+        $list->setColumnLayout('delete', ['', '<td class="rex-table-action">###VALUE###</td>']);
+        $list->setColumnParams('delete', ['func' => 'delete', 'wildcard_id' => '###id###']);
+        $list->addLinkAttribute('delete', 'data-confirm', $this->i18n('delete') . ' ?');
+    } else {
+        $list->addColumn('delete', '');
+        $list->setColumnLayout('delete', ['', '<td class="rex-table-action"></td>']);
+    }
+
 
     $content .= $list->get();
 
@@ -81,7 +106,7 @@ if ($func == '') {
     $form->setLanguageSupport('id', 'clang_id');
     $form->setEditMode($func == 'edit');
 
-    if ($func == 'add' && rex::getUser()->getComplexPerm('clang')->hasAll()) {
+    if (rex::getUser()->isAdmin()) {
         $field = $form->addTextField('wildcard', rex_request('wildcard_name', 'string', null));
         $field->setNotice($this->i18n('wildcard_without_tag'));
     } else {
